Seperated functionality to run modules seperatly
