+++ dockerfile.dockerfile
@@ -1,0 +1,9 @@
+FROM python:3-slim
+
+WORKDIR /usr/src/app 
+
+COPY requirement.txt ./
+RUN pip install --no-cache-dir -r requirements.txt
+
+COPY . .
+CMD ["python", "./main.py"]
