+++ requirement_all.txt
@@ -1,0 +1,118 @@
+
+absl-py==0.11.0
+argon2-cffi==20.1.0
+astunparse==1.6.3
+async-generator==1.10
+attrs==20.3.0
+backcall==0.2.0
+beautifulsoup4==4.9.3
+bleach==3.3.0
+cached-property==1.5.2
+cachetools==4.2.0
+certifi==2020.12.5
+cffi==1.14.5
+chardet==4.0.0
+click==7.1.2
+### cmake==3.18.2.post1  Creating error 
+###### dlib==19.18.0
+colorama==0.4.4
+comtypes==1.1.9
+DateTime==4.3
+decorator==4.4.2
+defusedxml==0.7.1
+entrypoints==0.3
+face-recognition==1.3.0
+face-recognition-models==0.3.0
+flatbuffers==1.12
+gast==0.4.0
+google-auth==1.24.0
+google-auth-oauthlib==0.4.2
+google-pasta==0.2.0
+grpcio==1.34.1
+h5py==3.1.0
+idna==2.10
+imageio==2.9.0
+imageio-ffmpeg==0.4.2
+importlib-metadata==3.3.0
+ipykernel==5.5.3
+ipython==7.16.1
+ipython-genutils==0.2.0
+ipywidgets==7.6.3
+jedi==0.18.0
+Jinja2==2.11.3
+jsonschema==3.2.0
+Keras==2.4.3
+keras-nightly==2.5.0.dev2021032900
+Keras-Preprocessing==1.1.2
+Markdown==3.3.3
+MarkupSafe==1.1.1
+mistune==0.8.4
+moviepy==1.0.3
+nbclient==0.5.3
+nbconvert==6.0.7
+nbformat==5.1.3
+nest-asyncio==1.5.1
+notebook==6.3.0
+numpy==1.19.4
+oauthlib==3.1.0
+opencv-python==4.4.0.46
+opt-einsum==3.3.0
+packaging==20.9
+pandocfilters==1.4.3
+parso==0.8.2
+pickleshare==0.7.5
+Pillow==8.0.1
+proglog==0.1.9
+prometheus-client==0.10.1
+prompt-toolkit==3.0.18
+protobuf==3.14.0
+py4j==0.10.9
+pyasn1==0.4.8
+pyasn1-modules==0.2.8
+PyAudio           Failed To build wheel for pyaudio $$$$$$$$ Error
+pycparser==2.20
+Pygments==2.8.1
+pyparsing==2.4.7
+pypiwin32
+pyrsistent==0.17.3
+python-dateutil==2.8.1
+pyttsx3==2.90
+pytz==2020.5
+pywin32==300
+pywinpty==0.5.7
+PyYAML==5.4
+pyzmq==22.0.3
+qtconsole==5.0.3
+QtPy==1.9.0
+requests==2.25.1
+requests-oauthlib==1.3.0
+rsa==4.6
+scipy==1.5.4
+Send2Trash==1.5.0
+six==1.15.0
+soupsieve==2.2.1
+SpeechRecognition==3.8.1
+tensorboard==2.5.0
+tensorboard-data-server==0.6.1
+tensorboard-plugin-wit==1.7.0
+tensorflow==2.5.0
+tensorflow-estimator==2.5.0
+termcolor==1.1.0
+terminado==0.9.4
+testpath==0.4.4
+tornado==6.1
+tqdm==4.51.0
+traitlets==4.3.3
+typing-extensions==3.7.4.3
+urllib3==1.26.4
+wcwidth==0.2.5
+webencodings==0.5.1
+Werkzeug==1.0.1
+widgetsnbextension==3.5.1
+wikipedia==1.4.0
+wincertstore==0.2
+wrapt==1.12.1
+zipp==3.4.0
+zope.interface==5.2.0
+
+     CMake must be installed to build the following extensions: dlib ERORRRR
