+++ requirement_min.txt
@@ -10,7 +10,7 @@ idna==2.10
 imageio==2.9.0
 imageio-ffmpeg==0.4.2
 moviepy==1.0.3
-numpy==1.18.4
+numpy>=1.18.4
 opencv-python==4.4.0.46
 Pillow==8.0.1
 proglog==0.1.9
@@ -18,10 +18,10 @@ requests==2.24.0
 tqdm==4.51.0
 urllib3==1.25.11
 wincertstore==0.2
-pyaudio==0.2
 SpeechRecognition==3.8.1
 pyttsx3==2.90
-pypiwin32==300
 wikipedia==1.4.0
 secure-smtplib==0.1.1
-
+pyaudio
+pypiwin32
+pyalsaaudio
