+++ settings.json
@@ -1,0 +1,5 @@
+{
+    "url":"https://www.flipkart.com/dell-ms-116-wired-optical-mouse/p/itmekm2ggehfhfaw?pid=ACCEKM2GZAGFW2TM&lid=LSTACCEKM2GZAGFW2TMF3S8CR&marketplace=FLIPKART&fm=organic&iid=ed685458-5a55-4487-b96c-eb1fba0ee480.ACCEKM2GZAGFW2TM.PRODUCTSUMMARY&ppt=pp&ppn=pp&ssid=dfgjpqwjxc0000001621429636046",
+    "budget":100,
+    "user-agent":"Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/90.0.4430.212 Safari/537.36" 
+}
